requirements.txt
cement==2.10.14
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.2
future==1.0.0
hpsklearn @ https://github.com/hyperopt/hyperopt-sklearn.git
hyperopt==0.2.7
idx2numpy==1.2.3
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.3
networkx==3.4.2
numpy==1.26.0
packaging==24.2
pandas==2.1.1
pillow==11.0.0
py4j==0.10.9.7
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.2
setuptools==75.6.0
six==1.17.0
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2024.2
